+++ __tests__/types.test.ts
@@ -518,13 +518,15 @@ describe('test built-in types', () => {
         .every((item: unknown) => !isNaN(item as number)),
     ).toBeTruthy();
     // with length
-    const numberAndString = Such.as(':::`:number`|`:string{5}`') as string;
+    const numberAndString = Such.as(
+      ':::`:number`|`:string:[95,122]:{5}`',
+    ) as string;
     expect(numberAndString.includes('|')).toBeTruthy();
     const numAndStrSegs = numberAndString.split('|');
     expect(numAndStrSegs.length >= 2).toBeTruthy();
-    const [num, ...strs] = numAndStrSegs;
+    const [num, str] = numAndStrSegs;
     expect(!isNaN((num as unknown) as number)).toBeTruthy();
-    expect(strs.join('').length).toEqual(5);
+    expect(str.length).toEqual(5);
   });
 });